Add some additional requirements checks in installer

+++ install.sh
@@ -13,12 +13,29 @@ if [ -f "$LOCAL_JAR_FILE" ]; then
 	cp -f $PWD/target/uberjar/atoss-cli-standalone.jar $INSTALL_DIR
 else
     echo "$LOCAL_JAR_FILE does not exist, pulling release from Github."
-	rm -f atoss-cli-standalone.jar
-	gh release download latest
-	mv -f atoss-cli-standalone.jar $INSTALL_DIR
+
+	if $(gh --version); then
+		# When Github CLI is installed
+		rm -f atoss-cli-standalone.jar
+		gh release download latest
+		mv -f atoss-cli-standalone.jar $INSTALL_DIR
+	else
+		echo "Github CLI is not set up! Please visit http://cli.github.com/ for more details."
+	fi
+
 fi
 
 echo "Installing wrapper"
 
 cp -f $PWD/wrapper.sh "${INSTALL_DIR}atoss-cli"
 
+# Make a check that chromedriver and java are present
+echo 'Checking requirements...'
+
+if ! $(java -version > /dev/null); then
+	echo "WARNING: Java is not installed or activated!"
+fi
+
+if ! $(chromedriver -v > /dev/null); then
+	echo "WARNING: chromedriver is not installed or is missing from PATH!"
+fi
